refactor: share paths between command and download
