+++ src/lib.rs
@@ -41,6 +41,7 @@ pub mod error;
 pub mod event;
 pub mod iter;
 pub mod log_parser;
+pub mod paths;
 pub mod pix_fmt;
 pub mod read_until_any;
 pub mod version;
